Modularize structure and decouple content
- Implemented modular directory structure: src/layouts, src/pages, and src/components.
- Decoupled configuration and text content into JSON files in the content/ directory.
- Refactored styling to use CSS variables in src/styles/tokens.css for easier theme swapping.
- Migrated all feature-based components to src/pages and shared hooks to src/hooks.
- Added MDX support and configured Vite for a standalone component library build.
- Fixed content loading issues by using recommended Vite glob import patterns.
- Updated pnpm-lock.yaml for CI consistency.
- Verified all changes via production build and lint checks.
