+++ src/App.tsx
@@ -12,7 +12,7 @@ import { Box } from './components/Primitives';
 import Home from './pages/Home';
 import GearReviews from './pages/Gear';
 import Research from './pages/Research';
-import Resources from './features/resources/ResourceGallery';
+import Resources from './pages/Resources';
 import About from './pages/About';
 import Blog from './pages/Blog';
 import BlogPost from './pages/BlogPost';
