+++ src/pages/ResearchDetail.tsx
@@ -2,7 +2,7 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { motion } from 'motion/react';
 import { Database, Activity, ArrowLeft, Search } from 'lucide-react';
 import { Box, Stack, Text, Grid } from '@/components/Primitives';
-import { useResearch } from '@/features/research/useResearch';
+import { useResearch } from '@/hooks/useResearch';
 import { BlogDrafter } from '@/components/BlogDrafter';
 
 export default function ResearchDetail() {
