+++ vite.config.ts
@@ -10,7 +10,9 @@ export default defineConfig(({mode}) => {
   return {
     plugins: [
       {
-        ...mdx(),
+        ...mdx({
+          include: /\.mdx$/
+        }),
         enforce: 'pre'
       },
       react(),
@@ -24,6 +26,25 @@ export default defineConfig(({mode}) => {
       alias: {
         '@': path.resolve(__dirname, './src'),
       },
+    },
+    build: {
+      lib: {
+        entry: path.resolve(__dirname, 'src/components/index.ts'),
+        name: 'TechDancerUI',
+        fileName: (format) => `tech-dancer-ui.${format}.js`
+      },
+      rollupOptions: {
+        external: ['react', 'react-dom', 'motion', 'lucide-react', 'react-router-dom'],
+        output: {
+          globals: {
+            react: 'React',
+            'react-dom': 'ReactDOM',
+            'motion': 'motion',
+            'lucide-react': 'LucideReact',
+            'react-router-dom': 'ReactRouterDOM'
+          }
+        }
+      }
     },
     server: {
       hmr: process.env.DISABLE_HMR ? false : {
